fix: resolve build warnings and action group registration error
- Replace deprecated GithubAuthenticationManager with GHAccountsUtil
- Remove redundant .toDouble() calls in StackGraphPanel
- Fix ChangesView.Toolbar → ChangesViewToolbar group ID

+++ src/main/kotlin/com/github/ydymovopenclawbot/stackworktree/ui/stackgraph/StackGraphPanel.kt
@@ -431,8 +431,8 @@ class StackGraphPanel : JPanel() {
         }
         val dotD = 8.0
         val dot  = Ellipse2D.Double(
-            (rightX - dotD).toDouble(),
-            (centerY - dotD / 2).toDouble(),
+            rightX - dotD,
+            centerY - dotD / 2,
             dotD,
             dotD,
         )
